Abstration of setCorrectDateText to eliminate repetition
